Feat/benchmark analyse result

+++ backend/benchmark/sentence_transformers.py
@@ -6,9 +6,9 @@ import pandas as pd
 from dotenv import load_dotenv
 from tools.elastic import get_index_name, replace_blue_green, create_index, put_alias, inject_documents, search, index_file, suggest, build_query
 import elasticsearch
-from elasticsearch import Elasticsearch
+from elasticsearch import Elasticsearch, helpers
 from benchmark.utils import *
-
+from sentence_transformers import SentenceTransformer
 
 '''Getting the arguments
 - the path of the Q/A test base
@@ -29,7 +29,9 @@ parser.add_argument('-m', dest='metric', type=str,
 parser.add_argument('-index',  action='store_true',
     help='Reindex?')
 
+
 def main(args):
+
     args.base_path = Path(args.base_path)
     #Loading the environment
     load_dotenv(dotenv_path=args.base_path / args.dotenv_path, override=True)
@@ -50,6 +52,10 @@ def main(args):
     ES_PORT=os.getenv('ES_PORT')
     ES_HOST=os.getenv('ES_HOST')
 
+    MODEL_NAME = "/app/benchmark/bld/data/sbert.net_models_" + \
+        "distilbert-multilingual-nli-stsb-quora-ranking"
+    # Embedding with Sentence transformers
+    model = SentenceTransformer(MODEL_NAME)
     # Section path
     path_sections = Path(USER_DATA) / 'sections.json'
     if path_sections.exists():
@@ -59,6 +65,7 @@ def main(args):
         sections = []
 
     os.makedirs(ES_DATA, exist_ok=True)
+
 
     #Reading files
     glossary_file = Path(USER_DATA) / GLOSSARY_FILE
@@ -72,55 +79,105 @@ def main(args):
         test_base_df = pd.read_csv(qr_path, encoding= 'utf-8') #if csv file
 
 
+    #Instanciation of elasticsearch
     es = Elasticsearch([{'host': ES_HOST, 'port': ES_PORT}])
-
     if args.index:
-        #Instanciation of elasticsearch
+        print("Index")
+        if es.indices.exists(index=INDEX_NAME):
+            for _ in range(3): # to be sure alias and indexes are removed
+                es.indices.delete(index=INDEX_NAME, ignore=[400, 404])
+                es.indices.delete_alias(index='_all',
+                    name=INDEX_NAME, ignore=[400, 404])
 
-        #Index creation
-        for i in range(3): # to be sure alias and indexes are removed
-            es.indices.delete(index=INDEX_NAME, ignore=[400, 404])
-            es.indices.delete_alias(index='_all',
-                name=INDEX_NAME, ignore=[400, 404])
-
-
+        # Index creation
         create_index(INDEX_NAME, USER_DATA, ES_DATA, MAPPING_FILE, GLOSSARY_FILE, RAW_EXPRESSION_FILE ) #some parameters are stored in the map, that is linked to the index here
-        #import pdb; pdb.set_trace()
-
 
 
         inject_documents(INDEX_NAME, USER_DATA, DST_DIR, JSON_DIR,
                         meta_path = META_DIR, sections=sections)
         time.sleep(1) # ! important, asynchronous injection
 
+
+        # Query all documents to retrieve questions
+        res = es.search(index=INDEX_NAME, body = {'_source': ['_id', '_type', 'question'],
+                'size' : 10000,
+                'query': {
+                    'match_all' : {}
+                }
+                })
+
+        all_questions={}
+        for r in res['hits']['hits']:
+            all_questions[r['_id']]= r['_source']['question'][0]
+
+        qids = list(all_questions.keys())
+        questions = [all_questions[qid] for qid in qids]
+
+        embeddings = model.encode(questions, show_progress_bar=False)
+
+        # Update the existing index with embedded vectors
+        for qid, embedding in zip(qids, embeddings):
+            es.update(index=INDEX_NAME,
+                doc_type='_doc',
+                id=qid,
+                body={
+                    "script" : {
+                        "source": "ctx._source.question_vector= params.emb",
+                        "lang": "painless",
+                        "params" : {
+                "emb" : embedding
+                }
+            }
+            })
+            import pdb; pdb.set_trace()
+
+        time.sleep(1)
+
     #Searching
     rank_body_requests = []
     rank_body_metric = metric_parameters(args.metric) # building the request metric
+
     query_dict = dict()
-    must = {}
-    should = {}
-    filter = ''
-    highlight = []
+
     #Building the request body
     for id, row in test_base_df.iterrows():
-
-        must = [{"multi_match":{"fields":["question","reponse","titre","mots-cles"],"query":row['Questions']}}]
-        body_built, length_of_request = build_query(must, should, filter, INDEX_NAME, highlight, glossary_file=glossary_file, expression_file=expression_file)
-        body_query = {"query":body_built["query"]} #removing highlight, to keep?
+        question_embedding = model.encode(row["Questions"])
+        #import pdb; pdb.set_trace()
+        body_query = {
+        "query": {
+            "script_score": {
+            "query": {
+                "match_all": {}
+            },
+            "script": {
+                "source": "cosineSimilarity(params.queryVector, doc['question_vector']) + 1.0",
+                "params": {
+                "queryVector": question_embedding
+                }
+            }
+            }
+        }
+        }
         request = { "id": str(id), "request": body_query, "ratings": [{ "_index": INDEX_NAME, "_id": "%s"%row['Fiches'], "rating": 1}]}
         rank_body_requests.append(request)
         query_dict[str(id)] = {"query": row['Questions'], "ratings": [{ "_index": INDEX_NAME, "_id": "%s"%row['Fiches'], "rating": 1}]}
+
+    #import pdb; pdb.set_trace()
     #Metric evaluation
     result = es.rank_eval(body= {
                                   "requests": rank_body_requests,
                                   "metric": rank_body_metric
                                   }, index = INDEX_NAME )
-
+    print(20*'-')
+    print('Result of %s script'%Path(__file__).resolve().stem)
+    print(20*'-')
+    #print(json.dumps(result,sort_keys=True, indent=4))
+    #import pdb; pdb.set_trace()
     #print(result)
     for id, req in result["details"].items():
         print("{question} => {answser}".format(question=query_dict[id]["query"],
                         answser=query_dict[id]["ratings"][0]["_id"]))
-        print("\n".join([hits['hit']['_id'] for hits  in req['hits']]))
+        print("\n".join(["%s (%.02f)"%(hits['hit']['_id'], hits['hit']['_score']) for hits  in req['hits']]))
         print("Score : %s"%req["metric_score"])
         print("\n")
